Sync merge trunk -> parallel (rev. 461-467)
-----

Send a warning when negative variances are set to zero.

* core/stk_predict.m: modified.

Overloaded quantile() function for dataframe objects.

* utils/@stk_dataframe/quantile.m: New function.

Improve backward compatibility.

* utils/@stk_dataframe/private/get_column_indicator.m: Return the
 'var' field if it exists when 'v' is asked for, and display a warning.

Fix a bug in stk_param_getdefaultbounds (not handling .a
structures properly).

* paramestim/stk_param_getdefaultbounds.m: Cast to double before
 calling var() to handle legacy .a structure properly.

+++ paramestim/stk_param_getdefaultbounds.m
@@ -40,7 +40,7 @@ else
     TOLSCALE = 5.0;
     
     % bounds for the variance parameter
-    empirical_variance = var(yi);
+    empirical_variance = var(double(yi));
     logvar_lb = min(log(empirical_variance), param0(1)) - TOLVAR;
     logvar_ub = max(log(empirical_variance), param0(1)) + TOLVAR;
